refactor: extract Places API fetch logic into useRestaurantSearch hook
